remove api types folder
