+++ src/pages/DashboardPage/components/GraceList/GraceList.tsx
@@ -1,7 +1,7 @@
 import { FC } from 'react';
 import { NavLink } from 'react-router-dom';
 
-import { GraceListItem } from '@frakt/api/raffle';
+import { GraceListItem } from '@frakt/state/liquidations/types';
 import Button from '../../../../components/Button';
 import { createTimerJSX } from '@frakt/utils';
 import { PATHS } from '../../../../constants';
