+++ src/pages/LiquidationsPage/components/OngoingRaffleTab/OngoingRaffleTab.tsx
@@ -10,8 +10,8 @@ import { useRaffleInfo } from '@frakt/hooks/useRaffleData';
 import EmptyList from '@frakt/componentsNew/EmptyList';
 import { useRaffleSort } from '../Liquidations/hooks';
 import styles from './OngoingRaffleTab.module.scss';
+import { RaffleListItem } from '@frakt/state/liquidations/types';
 import RafflesList from '../RafflesList';
-import { RaffleListItem } from '@frakt/api/raffle';
 import { useLiquidationRaffles } from './hooks';
 
 const OngoingRaffleTab: FC = () => {
