+++ src/state/liquidations/types.ts
@@ -19,6 +19,35 @@ export interface RaffleListItem {
   expiredAt: string;
 }
 
+export interface GraceListItem {
+  pubkey: string;
+  nftMint: string;
+  nftName: string;
+  nftImageUrl: string;
+  nftCollectionName: string;
+  isPriceBased: boolean;
+  loanValue: number;
+  liquidationPrice: number;
+  liquidationPriceLamports: string;
+  valuation: number;
+  startedAt: string;
+  expiredAt: string;
+}
+
+export interface WonRaffleListItem {
+  rafflePubKey: string;
+  nftMint: string;
+  nftName: string;
+  nftImageUrl: string;
+  nftCollectionName: string;
+  nftFloorPrice: number; //? 1.23 (SOL)
+  liquidationPrice: number; //? 1.23 (SOL)
+  liquidationPriceBN: string;
+  user: string;
+  expiredAt: string;
+  startedAt: string;
+}
+
 export interface LotteryTicket {
   totalTickets: number;
   currentTickets: number;
